ATIVA/requirementsFront.txt
pandas
plotly.graph_objects
dash


